1:Add unit test of user login model;2:Optimize the code of user login;3:Move loginUserInfo class from domain to vo package;
